update custom LR schedulers
